refactor: create models folder
