+++ src/lib.rs
@@ -5,6 +5,7 @@ extern crate dotenv;
 pub mod models;
 pub mod schema;
 
+use crate::models::*;
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
 use dotenv::dotenv;
@@ -19,8 +20,8 @@ pub fn establish_connection() -> PgConnection {
         .expect(&format!("Error connection to {}", database_url))
 }
 
-fn insert_test_data(conn: &PgConnection) -> models::Manga {
-    let manga = models::NewManga {
+fn insert_test_data(conn: &PgConnection) -> manga::Manga {
+    let manga = manga::NewManga {
         anilist_id:   1,
         cover_image:  "some bullshit".to_string(),
         banner_image: "another_bullshit".to_string(),
@@ -30,14 +31,14 @@ fn insert_test_data(conn: &PgConnection) -> models::Manga {
         title:        "ehe".to_string(),
     };
 
-    let staff = models::NewStaff {
+    let staff = staff::NewStaff {
         anilist_id: 1,
         image:      "some-link.com".to_string(),
         name:       "Onez".to_string(),
         role:       "Author".to_string(),
     };
 
-    let relation = models::NewRelation {
+    let relation = relation::NewRelation {
         media_type:        "Anime".to_string(),
         anilist_id:        111,
         status:            "On-going".to_string(),
@@ -46,29 +47,29 @@ fn insert_test_data(conn: &PgConnection) -> models::Manga {
         banner_image:      "some-link-again.com".to_string(),
     };
 
-    let m: models::Manga = diesel::insert_into(schema::mangas::table)
+    let m: manga::Manga = diesel::insert_into(schema::mangas::table)
         .values(&manga)
         .get_result(conn)
         .expect("Error cannot insert manga");
 
-    let r: models::Relation = diesel::insert_into(schema::relations::table)
+    let r: relation::Relation = diesel::insert_into(schema::relations::table)
         .values(&relation)
         .get_result(conn)
         .expect("Error cannot insert relation");
 
-    let s: models::Staff = diesel::insert_into(schema::staffs::table)
+    let s: staff::Staff = diesel::insert_into(schema::staffs::table)
         .values(&staff)
         .get_result(conn)
         .expect("Error cannot insert relation");
 
     // Do the association !! This is the crucial part.
 
-    let m_r = models::NewMedia {
+    let m_r = media::NewMedia {
         manga_id:    m.id,
         relation_id: r.id,
     };
 
-    let m_s = models::NewSeries {
+    let m_s = series::NewSeries {
         manga_id: m.id,
         staff_id: s.id,
     };
@@ -87,16 +88,16 @@ fn insert_test_data(conn: &PgConnection) -> models::Manga {
 }
 
 fn get_full_relation(
-    manga: &models::Manga,
+    manga: &manga::Manga,
     conn: &PgConnection,
-) -> std::vec::Vec<models::Staff> {
+) -> std::vec::Vec<staff::Staff> {
     use diesel::pg::expression::dsl::any;
 
     let manga_staff =
-        models::Series::belonging_to(manga).select(schema::series::staff_id);
+        series::Series::belonging_to(manga).select(schema::series::staff_id);
     schema::staffs::table
         .filter(schema::staffs::id.eq(any(manga_staff)))
-        .load::<models::Staff>(conn)
+        .load::<staff::Staff>(conn)
         .expect("Could not load tags")
 }
