+++ src/models/mod.rs
@@ -1,0 +1,5 @@
+pub mod manga;
+pub mod media;
+pub mod relation;
+pub mod series;
+pub mod staff;
